requirements.txt
# Các thư viện cơ bản
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2

# Web development (Flask/Django)
flask==2.3.2
flask_cors==6.0.0
django==4.2.3
requests==2.31.0

# Data science & Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
pytorch==2.0.1

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
mysql-connector-python==8.0.33

# Utilities
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.10.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0